FireDb: Bruge nedarvning fremfor imports
Refaktorisering med henblik på at hjælpe VSC og andre værktøjer med at
finde rundt i koden selvom den er delt ud på flere forskellige filer.
Den lidt ukonventionelle løsning med imports af funktioner ind i selve
klassen gør det svært for diverse værktøjer at regne ud hvor en given
funktion kommer fra. Det problem er løst ved i stedet at nedarve fra en
række klasser der alle har rod i BaseFireDb. Det vil sige vi får en
nedarvning i stil med

                      BaseFireDb

    FireDbHent      FireDbIndset      FireDbLuk

                       FireDb

I takt med at der kommer mere funktionalitet i FireDb kan det flyttes ud
i flere forskellige filer og klasser.
